Added CellState enum, small fixes, now works
