+++ src/Point.java
@@ -1,10 +1,8 @@
-package com.model;
-
 public class Point {
     public int x;
     public int y;
 
-    public Point(int x, int y) {
+    public Point(int x, int y){
         this.x = x;
         this.y = y;
     }
