+++ src/cell/CellState.java
@@ -1,0 +1,7 @@
+package cell;
+
+public enum CellState {
+    ALIVE,
+    DEAD
+}
+
